Add MinGW x86-64 Windows cross-build support
Extend the portable root Makefile with an OS=mingw branch (WIN32, SHLIB
for PE DLLs --export-all-symbols, EXE_SUFFIX=.exe, -lrt/-ldl gated off)
and make the Windows x86-64 port compile via x86_64-w64-mingw32-g++;
strongtalk.exe/stest.exe plus their PE DLLs are produced in
build/x86_64-mingw-gcc/.

Windows/LLP64 fixes across the shared sources (long is 32-bit on Windows):
rSet.hpp byte indexing, proxyOop long_at slots, prim.cpp RBX save/restore
thunks, callBack.cpp rel32 math and pointer prints -> intptr_t/%#tx;
std_includes.hpp no longer redefines __stdcall under mingw; types.hpp drops
the WIN32 int32_t typedefs in favor of inttypes.h; easyunit gains long long
StringFrom overloads.

os_nt.cpp is rewritten for x64/MinGW: replace the incls umbrella with the
per-header includes, read the stack pointer via __asm__ in startThread,
use DWORD ExceptionCode/exception_name, x64 CONTEXT Rsp/Rbp/Rip fields and
void** fp for the stack-overflow suspend hook, GetThreadTimes on the
thread handle, ExitProcess for fatalExit, DWORD_PTR affinity masks, and
jit_write_protect stubs.

Move the easyunit test harness (TestDeltaProcess, scheduler/event glue)
out of test/main/main.cpp into test/runtime/testProcess.{hpp,cpp} so the
PE test DLL is self-contained; PE shared libraries cannot carry undefined
symbols, so the Makefile links stest.so against strongtalk.so.

All four configs build with zero errors/warnings: macOS arm64, forced
macOS x86-64, Docker Linux/amd64, and the MinGW Windows cross-build.

+++ vm/memory/rSet.hpp
@@ -49,7 +49,7 @@ private:
 
   // friend void oldSpace::switch_pointers_by_card(oop, oop);
   char* byte_for(void* p) const {
-    return (char*)&byte_map[(((unsigned long)p) >> card_shift) - (((unsigned long)low_boundary) >> card_shift)];
+    return (char*)&byte_map[(((uintptr_t)p) >> card_shift) - (((uintptr_t)low_boundary) >> card_shift)];
   }
   //  char* byte_for(void *p) const { return (char*)&byte_map[int((char*)p - low_boundary) >> card_shift]; }
   oop* oop_for(char* p) const { return (oop*)(low_boundary + ((p - byte_map) << card_shift)); }
